+++ src/core/fs/read/reading.rs
@@ -1,7 +1,7 @@
 
 use crate::core::prelude::*;
 
-treatment!(file_reader_treatment,
+treatment!(reading_treatment,
     core_identifier!("fs","read";"Reading"),
     models![
         ("reader", crate::core::fs::read::files_reader::FileReaderModel::descriptor())
@@ -35,7 +35,7 @@ treatment!(file_reader_treatment,
             ok_or_break!(o_data.send_multiple_byte(bytes).await);
         }
 
-        if let (Ok(failure), Ok(messages)) = futures::join!(i_failure.recv_one_void(), i_message.recv_string()) {
+        if let (Ok(_), Ok(messages)) = futures::join!(i_failure.recv_one_void(), i_message.recv_string()) {
 
             let _ = o_failure.send_void(()).await;
             let _ = o_message.send_multiple_string(messages).await;